zerocopy_derive::derive

Module project

Enums

Validity 🔒

Functions

derive 🔒

derive_enum 🔒

derive_projection_struct_union 🔒

Generates field projection implementations for a struct or union.

generate_tag_consts 🔒

Generates a constant for the tag associated with each variant of the enum. When we match on the enum’s tag, each arm matches one of these constants. We have to use constants here because:

struct_union_variant_id 🔒

Returns the projection variant ID for the struct or union described by ctx.

tag_ident 🔒

variant_struct_ident 🔒

variants_union_field_ident 🔒