zerocopy_derive::derive

Module try_from_bytes

Functions

derive_has_field_struct_union 🔒

derive_is_bit_valid 🔒

Generates an implementation of is_bit_valid for an arbitrary enum.

derive_try_from_bytes 🔒

derive_try_from_bytes_enum 🔒

derive_try_from_bytes_struct 🔒

derive_try_from_bytes_union 🔒

gen_trivial_is_bit_valid_unchecked 🔒 ⚠

Safety

generate_tag_consts 🔒

Generates a constant for the tag associated with each variant of the enum. When we match on the enum’s tag, each arm matches one of these constants. We have to use constants here because:

generate_variant_structs 🔒

Generates variant structs for the given enum variant.

generate_variants_union 🔒

tag_ident 🔒

try_gen_trivial_is_bit_valid 🔒

variant_struct_ident 🔒

variants_union_field_ident 🔒