zerocopy_derive

Function gen_trivial_is_bit_valid_unchecked

pub(crate) unsafe fn gen_trivial_is_bit_valid_unchecked(
    zerocopy_crate: &Path,
) -> TokenStream

Generates a TryFromBytes::is_bit_valid instance that unconditionally returns true.

This should be used where possible, (although try_gen_trivial_is_bit_valid should be preferred over this for safety reasons). Using this impl is faster to codegen, faster to compile, and is friendlier on the optimizer.

Safety

The caller must ensure that all initialized bit patterns are valid for Self.