zerocopy_derive

Module repr

Modules

util 🔒

Structs

UnrecognizedReprError 🔒

The representation hint could not be parsed or was unrecognized.

UnsupportedReprError 🔒

The representation hint is not supported for the decorated type.

Enums

AlignRepr 🔒

repr(packed(...)) or repr(align(...))

CompoundRepr 🔒

A compound representation: repr(C), repr(Rust), or repr(Int).

FromAttrsError 🔒

The error returned from Repr::from_attrs.

FromRawReprError 🔒

The error from converting from a RawRepr.

FromRawReprsError 🔒

The error from extracting a high-level repr type from a list of RawReprs.

PrimitiveRepr 🔒

repr(Int)

RawRepr 🔒

The result of parsing a single #[repr(...)] attribute or a single directive inside a compound #[repr(..., ...)] attribute.

Repr 🔒

The computed representation of a type.

Functions

try_from_raw_reprs 🔒

Tries to extract a high-level repr from a list of RawReprs.

Type Aliases

EnumRepr 🔒

The representations which can legally appear on an enum type.

StructUnionRepr 🔒

The representations which can legally appear on a struct or union type.