zerocopy_derive::repr

Enum Repr

pub(crate) enum Repr<Prim, Packed> {
    Transparent(Span),
    Compound(Spanned<CompoundRepr<Prim>>, Option<Spanned<AlignRepr<Packed>>>),
}

The computed representation of a type.

This is the result of processing all #[repr(...)] attributes on a type, if any. A Repr is only capable of representing legal combinations of #[repr(...)] attributes.

Variants

Transparent(Span)

#[repr(transparent)]

Compound(Spanned<CompoundRepr<Prim>>, Option<Spanned<AlignRepr<Packed>>>)

A compound representation: repr(C), repr(Rust), or repr(Int) optionally combined with repr(packed(...)) or repr(align(...))

Implementations

impl<Prim, Packed> Repr<Prim, Packed>

pub(crate) fn repr_type_name(&self) -> &str
where Prim: Copy + With<PrimitiveRepr>,

Gets the name of this “repr type” - the non-align repr(X) that is used in prose to refer to this type.

For example, we would refer to #[repr(C, align(4))] struct Foo { ... } as a “repr(C) struct”.

pub(crate) fn is_transparent(&self) -> bool

pub(crate) fn is_c(&self) -> bool

pub(crate) fn is_primitive(&self) -> bool

pub(crate) fn get_packed(&self) -> Option<&Packed>

pub(crate) fn get_align(&self) -> Option<Spanned<NonZeroU32>>

pub(crate) fn is_align_gt_1(&self) -> bool

pub(crate) fn unaligned_validate_no_align_gt_1(&self) -> Result<(), Error>

When deriving Unaligned, validate that the decorated type has no #[repr(align(N))] attribute where N > 1. If no such attribute exists (including if N == 1), this returns Ok(()), and otherwise it returns a descriptive error.

impl<Prim> Repr<Prim, NonZeroU32>

pub(crate) fn is_packed_1(&self) -> bool

Does self describe a #[repr(packed)] or #[repr(packed(1))] type?

impl<Packed> Repr<PrimitiveRepr, Packed>

fn get_primitive(&self) -> Option<&PrimitiveRepr>

pub(crate) fn is_u8(&self) -> bool

Does self describe a #[repr(u8)] type?

pub(crate) fn is_i8(&self) -> bool

Does self describe a #[repr(i8)] type?

impl<Prim, Packed> Repr<Prim, Packed>

fn from_attrs_inner( attrs: &[Attribute], ) -> Result<Repr<Prim, Packed>, Spanned<FromAttrsError>>
where Prim: With<PrimitiveRepr>, Packed: With<NonZeroU32>,

impl<Prim, Packed> Repr<Prim, Packed>

pub(crate) fn from_attrs( attrs: &[Attribute], ) -> Result<Repr<Prim, Packed>, Error>
where Prim: With<PrimitiveRepr>, Packed: With<NonZeroU32>,